Split transport/nats file to match transport/rpc format
