+++ rpc/transport/nats/server.go
@@ -1,9 +1,6 @@
 package nats
 
 import (
-	"fmt"
-	"time"
-
 	"github.com/nats-io/nats-server/v2/server"
 	"github.com/nats-io/nats.go"
 	"github.com/rs/zerolog/log"
@@ -12,16 +9,12 @@ import (
 const (
 	nitroRequestTopic      = "nitro-request"
 	nitroNotificationTopic = "nitro-notify"
+	apiVersionPath         = "/api/v1"
 )
 
 type natsTransport struct {
 	nc                *nats.Conn
 	natsSubscriptions []*nats.Subscription
-}
-
-type natsTransportClient struct {
-	natsTransport
-	notificationChan chan []byte
 }
 
 type natsTransportServer struct {
@@ -38,6 +31,24 @@ func newNatsTransport(url string) (*natsTransport, error) {
 		nc:                nc,
 		natsSubscriptions: make([]*nats.Subscription, 0),
 	}, nil
+}
+
+func (c *natsTransport) Close() {
+	for _, sub := range c.natsSubscriptions {
+		err := c.unsubscribeFromTopic(sub, 0)
+		if err != nil {
+			log.Error().Err(err).Msgf("failed to unsubscribe from a topic: %s", sub.Subject)
+		}
+	}
+	c.nc.Close()
+}
+
+func (c *natsTransport) unsubscribeFromTopic(sub *nats.Subscription, try int32) error {
+	err := sub.Unsubscribe()
+	if err != nil && try < 3 {
+		return c.unsubscribeFromTopic(sub, try+1)
+	}
+	return nil
 }
 
 func NewNatsTransportAsServer(rpcPort int) (*natsTransportServer, error) {
@@ -60,26 +71,8 @@ func NewNatsTransportAsServer(rpcPort int) (*natsTransportServer, error) {
 	return con, nil
 }
 
-func NewNatsTransportAsClient(url string) (*natsTransportClient, error) {
-	natsTransport, err := newNatsTransport(url)
-	if err != nil {
-		return nil, err
-	}
-	return &natsTransportClient{
-		natsTransport: *natsTransport,
-	}, nil
-}
-
-func (c *natsTransportClient) Request(data []byte) ([]byte, error) {
-	msg, err := c.nc.Request(nitroRequestTopic, data, 10*time.Second)
-	if msg == nil {
-		return nil, fmt.Errorf("received nill data for request %v with error %w", data, err)
-	}
-	return msg.Data, err
-}
-
 func (c *natsTransportServer) RegisterRequestHandler(apiVersion string, handler func([]byte) []byte) error {
-	sub, err := c.nc.Subscribe(nitroRequestTopic, func(msg *nats.Msg) {
+	sub, err := c.nc.Subscribe(nitroRequestTopic+"/api/"+apiVersion, func(msg *nats.Msg) {
 		responseData := handler(msg.Data)
 		err := c.nc.Publish(msg.Reply, responseData)
 		if err != nil {
@@ -89,24 +82,6 @@ func (c *natsTransportServer) RegisterRequestHandler(apiVersion string, handler 
 	c.natsSubscriptions = append(c.natsSubscriptions, sub)
 
 	return err
-}
-
-func (c *natsTransportClient) Subscribe() (<-chan []byte, error) {
-	if c.notificationChan != nil {
-		return c.notificationChan, nil
-	}
-	c.notificationChan = make(chan []byte)
-	subscription, err := c.nc.Subscribe(nitroNotificationTopic, func(msg *nats.Msg) {
-		c.notificationChan <- msg.Data
-	})
-	c.natsSubscriptions = append(c.natsSubscriptions, subscription)
-
-	return c.notificationChan, err
-}
-
-func (c *natsTransportClient) Close() {
-	c.natsTransport.Close()
-	close(c.notificationChan)
 }
 
 func (c *natsTransportServer) Notify(data []byte) error {
@@ -120,22 +95,4 @@ func (c *natsTransportServer) Url() string {
 func (c *natsTransportServer) Close() {
 	c.natsTransport.Close()
 	c.ns.Shutdown()
-}
-
-func (c *natsTransport) Close() {
-	for _, sub := range c.natsSubscriptions {
-		err := c.unsubscribeFromTopic(sub, 0)
-		if err != nil {
-			log.Error().Err(err).Msgf("failed to unsubscribe from a topic: %s", sub.Subject)
-		}
-	}
-	c.nc.Close()
-}
-
-func (c *natsTransport) unsubscribeFromTopic(sub *nats.Subscription, try int32) error {
-	err := sub.Unsubscribe()
-	if err != nil && try < 3 {
-		return c.unsubscribeFromTopic(sub, try+1)
-	}
-	return nil
 }
